services/auth_service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
pydantic==2.7.1
PyJWT==2.8.0
